requirements.txt
asgiref==3.6.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.2
fonttools==4.39.4
html5lib==1.1
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
social-auth-app-django
django-simple-history~=3.3.0